Firma de payu para recibir pagos certificados

+++ src/main/java/com/wow/libre/domain/CreatePaymentRedirectDto.java
@@ -4,6 +4,8 @@ import com.wow.libre.domain.enums.*;
 import com.wow.libre.domain.model.*;
 import lombok.*;
 
+import java.math.*;
+
 @Builder
 public class CreatePaymentRedirectDto {
     public String redirect;
@@ -13,7 +15,7 @@ public class CreatePaymentRedirectDto {
     public String currency;
     public String taxReturnBase;
     public String tax;
-    public Double amount;
+    public BigDecimal amount;
     public String referenceCode;
     public String description;
     public boolean isPayment;
